Make index cursors take longs to allow for the full unsigned 32 bit range. Implements [5ec9c66b0cf42].

+++ io7m-jcanephora-tests/src/test/java/com/io7m/jcanephora/tests/contracts/IndexBufferMapContract.java
@@ -260,7 +260,7 @@ import com.io7m.junreachable.UnreachableCodeException;
       final IndexBufferReadMappedType map = gm.indexBufferMapRead(ib);
       final CursorReadableIndexType cursor = map.getCursor();
       for (int index = 0; index < ib.bufferGetRange().getInterval(); ++index) {
-        final int value = cursor.getIndex();
+        final long value = cursor.getIndex();
         Assert.assertTrue(value == index);
       }
     }
@@ -304,7 +304,7 @@ import com.io7m.junreachable.UnreachableCodeException;
       final IndexBufferReadMappedType map = gm.indexBufferMapRead(ib);
       final CursorReadableIndexType cursor = map.getCursor();
       for (int index = 0; index < ib.bufferGetRange().getInterval(); ++index) {
-        final int value = cursor.getIndex();
+        final long value = cursor.getIndex();
         Assert.assertTrue(value == index);
       }
     }
@@ -347,7 +347,7 @@ import com.io7m.junreachable.UnreachableCodeException;
       final IndexBufferReadType map = gm.indexBufferMapRead(ib);
       final CursorReadableIndexType cursor = map.getCursor();
       for (int index = 0; index < ib.bufferGetRange().getInterval(); ++index) {
-        final int value = cursor.getIndex();
+        final long value = cursor.getIndex();
         Assert.assertTrue(value == index);
       }
     }
